Add bootstrap, use html parser, modify fetcher
Added bootstrap to determine if Golang is installed correctly on
the machine and if so gets the needed html parser from golang.org/x/net/html.

Using an html parser to easily grab the hyperlinks that are contained within
the page.

Modified the fetcher interface to not return the body which is the title of
the page. Currently do not want this as a feature of the interface and can be
easily added later down the road.