Revamp the entire crawler
Docker-compose:
Added the ability to use docker compose so that we can send all urls to a redis queue
based on their hostname. We build a docker image so that we can then link it to the
redis container.

All urls crawled will now be stored in a redis server in host-sharded queues so that we
can then later have a indexing application go through each of the queues and perform indexing
and then store the results to a SQL database.

I prefer the design of this since it will not need to do anything more than visit the site
parse the html and crawl the url it finds on the page. Going to be adding a caching
mechanism to the program so that we are not visiting a page more than once.