Cache urls, add concurrent Crawling
Implemented concurrent crawling of urls so that we can speed up the process of crawling the
website that we wnat to visit. After running this program for 2 hours last night it was still
crawling even with a depth of 5 so lets speed this up.

Implemented a caching mechanisms so that concurrent go routines will not be able to crawl pages
that other routines are working on thus reducing overhead and repetitive work.